+++ sql_template/l1/fact_stock_holder_log.sql
@@ -6,10 +6,10 @@ create table if not exists l1.fact_stock_holder_log (
     holder_num      bigint      comment '股东户数',
     holder_ann_date string      comment '股东户数公告日期',
     holder_end_date string      comment '股东户数截止日期',
-    total_mv        float       comment '最新市值(亿)', -- 最新市值
-    circ_mv         float       comment '流通市值(亿)', -- 最新流通市值
+    total_mv        double       comment '最新市值(亿)', -- 最新市值
+    circ_mv         double       comment '流通市值(亿)', -- 最新流通市值
 --     per_amount      float       comment '流通户均持股金额（万）=流通市值/持股人数',
-    `close`         float       comment '最新收盘价',
+    `close`         double       comment '最新收盘价',
     area            string      comment '地域',
     industry        string      comment '所属行业',
     market          string      comment '市场类型（主板/创业板/科创板/CDR）',
